Use mustache to render static record list values
